requirements.txt
iso639-lang==2.0.0
langdetect==1.0.9
numpy==1.21.2
Pillow==8.3.2
pysword==0.2.7
pywin32==301
six==1.16.0
wxPython==4.1.1
